chore: add tests to make sure strings are borrowed/owned
